refactor(desktop): extract context budget policy
